refactor : enum 클래스 위치를 constants 패키지로 변경
