+++ voucher/src/main/java/com/example/voucher/domain/FixedAmountVoucher.java
@@ -2,7 +2,7 @@ package com.example.voucher.domain;
 
 import java.util.UUID;
 
-import com.example.voucher.domain.enums.VoucherType;
+import com.example.voucher.constant.VoucherType;
 import com.example.voucher.utils.validator.VoucherValidator;
 
 public class FixedAmountVoucher implements Voucher {
