+++ voucher/src/main/java/com/example/voucher/domain/dto/VoucherDTO.java
@@ -1,8 +1,6 @@
 package com.example.voucher.domain.dto;
 
-import java.util.UUID;
-
-import com.example.voucher.domain.enums.VoucherType;
+import com.example.voucher.constant.VoucherType;
 
 public record VoucherDTO(long value, VoucherType voucherType) {
