+++ voucher/src/main/java/com/example/voucher/service/VoucherService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 import com.example.voucher.domain.FixedAmountVoucher;
 import com.example.voucher.domain.PercentDiscountVoucher;
 import com.example.voucher.domain.Voucher;
-import com.example.voucher.domain.enums.VoucherType;
+import com.example.voucher.constant.VoucherType;
 import com.example.voucher.repository.VoucherRepository;
 
 @Service
